Introduce bazel and abseil.

+++ src/ranker.cc
@@ -5,7 +5,7 @@
 int_t Ranker::GetKMerRank(const std::string& k_mer) const {
   int_t rank = 0;
   for (const char base : k_mer) {
-    rank = 4 * rank + base_rank[base];
+    rank = 4 * rank + base_rank[(int)base];
   }
   return rank;
 }
@@ -17,5 +17,5 @@ int_t Ranker::GetKMerRank(const int_t hint_rank, const char c) const {
   int_t rank = hint_rank & (power_of_k_ - 1);
 
   // Add last character contribution to the rank.
-  return 4 * rank + base_rank[c];
+  return 4 * rank + base_rank[(int)c];
 }
